Use nvFatBinSegment to find nv_fatbin offsets

+++ src/hardware/device/cuda/CUDAAccelerator.cpp
@@ -4,6 +4,7 @@
 	Copyright (C) 2020-2022 Barcelona Supercomputing Center (BSC)
 */
 
+#include <algorithm>
 #include <array>
 
 #include "CUDAAccelerator.hpp"
@@ -130,13 +131,13 @@ void CUDAAccelerator::callTaskBody(Task *task, nanos6_address_translation_entry_
 		// we COULD perform some math to express the same working units in a supported way.
 		// Right now we expect the user to provide valid parameters.
 
-		size_t gridDim1 = std::max(deviceInfo.sizes[0], 1);
-		size_t gridDim2 = std::max(deviceInfo.sizes[1], 1);
-		size_t gridDim3 = std::max(deviceInfo.sizes[2], 1);
+		size_t gridDim1 = std::max((int64_t) deviceInfo.sizes[0], (int64_t) 1);
+		size_t gridDim2 = std::max((int64_t) deviceInfo.sizes[1], (int64_t) 1);
+		size_t gridDim3 = std::max((int64_t) deviceInfo.sizes[2], (int64_t) 1);
 
-		size_t blockDim1 = std::max(deviceInfo.sizes[3], 1);
-		size_t blockDim2 = std::max(deviceInfo.sizes[4], 1);
-		size_t blockDim3 = std::max(deviceInfo.sizes[5], 1);
+		size_t blockDim1 = std::max((int64_t) deviceInfo.sizes[3], (int64_t) 1);
+		size_t blockDim2 = std::max((int64_t) deviceInfo.sizes[4], (int64_t) 1);
+		size_t blockDim3 = std::max((int64_t) deviceInfo.sizes[5], (int64_t) 1);
 
 		std::array<void *, MAX_STACK_ARGS> stack_params;
 		void **params = &stack_params[0];
